+++ handler/failover.go
@@ -17,15 +17,20 @@ type FailoverHandlerOptions struct {
 	ContinueOnError bool
 }
 
+// ContextWithFailoverHandlerOptions adds the options to the given context and returns the new context.
+func ContextWithFailoverHandlerOptions(ctx context.Context, opts FailoverHandlerOptions) context.Context {
+	return context.WithValue(ctx, failoverHandlerOptionsContext{}, &opts)
+}
+
 // DefaultFailoverHandlerOptions returns a default set of options for the handler.
 func DefaultFailoverHandlerOptions() FailoverHandlerOptions {
 	return FailoverHandlerOptions{}
 }
 
-// GetFailoverHandlerOptionsFromContext retrieves the options from the context.
+// FailoverHandlerOptionsFromContext retrieves the options from the context.
 //
 // If the options are not set in the context, a set of default options is returned instead.
-func GetFailoverHandlerOptionsFromContext(ctx context.Context) *FailoverHandlerOptions {
+func FailoverHandlerOptionsFromContext(ctx context.Context) *FailoverHandlerOptions {
 	o := ctx.Value(failoverHandlerOptionsContext{})
 	if o != nil {
 		if opts, ok := o.(*FailoverHandlerOptions); ok {
@@ -34,11 +39,6 @@ func GetFailoverHandlerOptionsFromContext(ctx context.Context) *FailoverHandlerO
 	}
 	opts := DefaultFailoverHandlerOptions()
 	return &opts
-}
-
-// AddToContext adds the options to the given context and returns the new context.
-func (o *FailoverHandlerOptions) AddToContext(ctx context.Context) context.Context {
-	return context.WithValue(ctx, failoverHandlerOptionsContext{}, o)
 }
 
 // failoverHandler simply sends the log message to the first available handler.
@@ -58,7 +58,7 @@ func NewFailoverHandler(opts FailoverHandlerOptions, handlers ...slog.Handler) *
 
 // Enabled determines whether or not the given level is enabled for any handler.
 func (h failoverHandler) Enabled(ctx context.Context, l slog.Level) bool {
-	handlerCtx := h.options.AddToContext(ctx)
+	handlerCtx := ContextWithFailoverHandlerOptions(ctx, h.options)
 	for _, handler := range h.handlers {
 		if handler.Enabled(handlerCtx, l) {
 			return true
@@ -70,7 +70,7 @@ func (h failoverHandler) Enabled(ctx context.Context, l slog.Level) bool {
 // Handle is responsible for finding the first available handler to write the record.
 func (h *failoverHandler) Handle(ctx context.Context, r slog.Record) error {
 	var err error
-	handlerCtx := h.options.AddToContext(ctx)
+	handlerCtx := ContextWithFailoverHandlerOptions(ctx, h.options)
 
 	for _, handler := range h.handlers {
 		if handler.Enabled(handlerCtx, r.Level) {